backend_backup/requirements.txt
# Essentials for the API
fastapi
uvicorn
pydantic
google-genai

# ML Essentials (CPU only)
torch
numpy
scikit-learn